Seperated controller and eventloop to make it more modulized
